Implement our own CorsFilter
This is because Chrome is stricter about the implementation.  See:

* https://stackoverflow.com/questions/10883211/deadly-cors-when-http-localhost-is-the-origin
* https://bugs.chromium.org/p/chromium/issues/detail?id=67743
* http://williambert.online/2013/06/allow-cors-with-localhost-in-chrome/

Personally, I think the issue is that the EBISPOT CorsFilter is
returning the actual resource with a status code 200 to a pre-flight
request, and that Chrome needs an actual 204 response when this is an
OPTIONS request.  The Express logic in the last link is essentially
implementing a middleware.