With the update to Spring 5.3, the extension ".json" no longer works.
The issue here is likely that an older version of Spring accepted any
text after "swagger" at the controller.  The current version is more
strict.  In order to provide a more timely fix, I am not worrying about
content negotiation or Cool URIs - that can come later.

This also patches the version of postcss used during the build process,
because yarn complained.