Update footer for HHS Vulnerability Disclosure link (#191)
* With the update to Spring 5.3, the extension ".json" no longer works.

The issue here is likely that an older version of Spring accepted any
text after "swagger" at the controller.  The current version is more
strict.  In order to provide a more timely fix, I am not worrying about
content negotiation or Cool URIs - that can come later.

This also patches the version of postcss used during the build process,
because yarn complained.

* Improve fix for swagger UI and bump version to 1.1.16

- Improve fix for swagger UI by convincing Spring to serve as
  /swagger/swagger.json
- Bump version to 1.1.16

* Bump version to test CI/CD.

* Update to add VDP footer link.

* Patch version of shelljs to keep dependabot happy.

* Move around the footer links a bit.

* Small tweak to footer for VDP.

* Bump version to 1.1.17